requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.7
requests==2.32.3
psycopg2==2.9.10
celery==5.4.0
pika==1.3.2
redis==5.2.1
requests==2.32.3